Make test run as part of the test suite instead of only alone.
